Can handle folders

+++ src/renderer/components/FileItem.vue
@@ -17,7 +17,7 @@
             <hy-button v-if="type == 'local'" @click="upload()" type="transparent"><i class="icon-upload"></i></hy-button>
             <hy-button v-if="type == 'remote'" @click="download()" type="transparent"><i class="icon-download"></i></hy-button>
 
-            <hy-button @click="deleteFile()" type="transparent"><i class="icon-trash"></i></hy-button>
+            <hy-button @click="deleteObject()" type="transparent"><i class="icon-trash"></i></hy-button>
 
             <hy-button @click="renaming.active = true" type="transparent"><i class="icon-pen"></i></hy-button>
         </div>
@@ -57,13 +57,26 @@ export default {
             deep: true,
         },
     },
+    computed: {
+        isDir() {
+            return this.file.type == "d";
+        },
+        isFile() {
+            return this.file.type == "-";
+        },
+        isLink() {
+            return this.file.type == "l";
+        },
+    },
     methods: {
         async upload() {
             // Remote path must include a valid filename
             let remoteFilePath = pathModule.join(this.tabInfo.remote.path, this.file.name);
 
             try {
-                await this.client.fastPut(this.file.path, remoteFilePath);
+                if (this.isFile) await this.client.fastPut(this.file.path, remoteFilePath);
+                else if (this.isDir) await this.client.uploadDir(this.file.path, remoteFilePath);
+
                 this.$emit("fetchRemote");
             } catch (error) {
                 console.error(error);
@@ -75,22 +88,27 @@ export default {
             let localFilePath = pathModule.join(this.tabInfo.local.path, this.file.name);
 
             try {
-                await this.client.fastGet(this.file.path, localFilePath);
+                if (this.isFile) await this.client.fastGet(this.file.path, localFilePath);
+                else if (this.isDir) await this.client.downloadDir(this.file.path, localFilePath);
                 this.$emit("fetchLocal");
             } catch (error) {
                 console.error(error);
                 alert("Error while downloading");
             }
         },
-        async deleteFile() {
+        async deleteObject() {
             if (!confirm(`Shure to delete '${this.file.name}' from ${this.type}?`)) return;
 
             try {
                 if (this.type == "remote") {
-                    await this.client.delete(this.file.path);
+                    if (this.isFile) await this.client.delete(this.file.path);
+                    else if (this.isDir) await this.client.rmdir(this.file.path, true);
+
                     this.$emit("fetchRemote");
                 } else {
-                    await fsp.rm(this.file.path);
+                    await fsp.rm(this.file.path, {
+                        recursive: true,
+                    });
                     this.$emit("fetchLocal");
                 }
             } catch (error) {
